mobiwiseinsight/requirements.txt
Flask
pyodbc
requests
apscheduler
flask-compress
flask-caching
python-jwt
pywhatkit
pygame
smtplib
python-decimal
email
uuid
logging
google-generativeai